DStarRepeater:  Shut down repeater thread nicely.
Delete the thread object after we shut it down in the OnExit() method of the
app.  This reduces the warning in the debug output that the thread hasn't
been shut down properly.  This will probably need to be done for the rest of
any threads that run in the app as well.