DStarRepeater:  Start move to using packed struct for parsing
Packed structs are easier to read when you're trying to get things parsed
out of fixed binary packet formats.  We use a struct defined at the top of
the file to parse gateway packets rather than picking through it as an array.
This also means we get to directly assign 16-bit and 32-bit values instead
of doing a lot of math to them merely to get the right bits in the right
places.  This patch converts to of the functions here to use the struct.
The other ones will come later.