ircDDBGateway:  Do hosts file lookups on background thread
The gateway tries to resolve all of the DNS addresses in the various hosts
files at startup.  If there is a DNS issue for some reason, this can take a
very long time.  Meanwhile the program is completely useless as it waits to
look up all these hosts.

This commit forks off three background threads to do the DNS resolution and
refrains from blocking the main thread of execution.  This way the program
can continue immediately.  The threads will destroy themselves upon exit.