ircDDBGateway:  Get rid of IRCDDBGatewayAppD
the wxUSE_GUI=0 preprocessor directive will make the application use a
console event loop rather than the GUI one.  We can use this to consolodate
both version of the application into one large source.  This gives code
commonality and it potentially allows for the *d apps to work on Windows
as well as UNIX.  This mirrors what has already been done for dstarrepeater.